pyproject.toml
[tool.poetry]
name = "testapp"
version = "0.1.0"
description = ""
authors = ["apoorv1125 <apoorv.gupta@mindstix.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
alembic = "1.16.5"
bcrypt = "5.0.0"
blinker = "1.9.0"
click = "8.1.8"
distlib = "0.4.0"
filelock = "3.19.1"
Flask = "3.1.2"
Flask-Bcrypt = "1.0.1"
Flask-Login = "0.6.3"
Flask-Migrate = "4.1.0"
Flask-SQLAlchemy = "3.1.1"
importlib-metadata = "8.7.0"
itsdangerous = "2.2.0"
Jinja2 = "3.1.6"
Mako = "1.3.10"
MarkupSafe = "3.0.3"
platformdirs = "4.4.0"
PyJWT = "2.10.1"
SQLAlchemy = "2.0.44"
tomli = "2.3.0"
typing-extensions = "4.15.0"
virtualenv = "20.35.4"
Werkzeug = "3.1.4"
zipp = "3.23.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
Flask-JWT-Extended = "^4.7.1"
marshmallow = "^4.1.2"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
